+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog of openradar
 0.3.8 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add basic nowcast product, task and script
+
+- Improve image script product selection
 
 
 0.3.7 (2013-08-29)